texture_cache: separate file for image_info
